requirements.txt
# Core deep learning framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Scientific computing and data handling
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization and plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# EEG-specific libraries
mne>=1.4.0

# Configuration management
PyYAML>=6.0

# Progress bars and utilities
tqdm>=4.65.0

# Image processing
Pillow>=9.5.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0

# File format support
h5py>=3.8.0

# Memory profiling
psutil>=5.9.0